Allow more freedom for FESOM mesh path
Before a mesh had to be in a directory called the same as the mesh name.
If mesh was in a folder 'mesh' then pickle files etc. would not work,
mesh name in mesh metadata would be 'mesh'...

Now the mesh name is inferred from the ENTIRE mesh path:
mesh name will be deepest directory name that is not "mesh"