Fix timechunking + add horizontal selections/mean
Small fixes regarding time_clounter to be given as argument. Also the
time_centered variable is chunked differently by default than the data.
But this leads to problems when trying to chunk new variables the same
way as other variables are chunked (misaligned chunk sizes). Thus only
loading or rechunking the the coordinate is a solution.

Add routines to make horizontal selection and averages. For now only
global averages and global selection are implemented