compute w_cos weights: Dont always convert to dask
Do not make the w_cos weights for global mean regular data always a dask
array. Only do this if data is not already fully loaded. Or if part of
the data is already a dask array.

This is also a fix that loaded data can be weighted. Otherwise this
would not be possible.

Alternativaly one could set chunks to None. This makes the weigths a
dask array, but does not rely on the dataset containing a dask array.