sub_oifs: Convert radiation from Jm-2 to W-m2
This is done automatically for only certain values and only if the
output frequency (accumulation period) is given in hours and only if the
data has the right metadata.

This should be a very safe implementation. It will not do anything if
these criteria are not fulfilled.

It might crash if data has not the correct metadata, but let's see. In
that case one would have to check first if the data has metadate. If it
doesn't it should also just return the orininal unconverted dataset like
in the other cases