Fix extended colorbars
Without this fix the over and under values of the colorbars are the same
as the last/first color. This results in there being do visible
difference for data that is outside the plot range (indicated by the
extend arrows on the colorbar). This fix is implemented only for linear
scaling as it might break for logarithmic scaling.

Implementation: In addition to the midpoits of the colorlevels that should
be plotted, two additional mid-layers are extrapolated. Every colorbar
has them. Thus for plots which do not make use of the extended colors
there will be two colors less in the palette, thus the color range will
be slightly shallower with this implementation.