open_data: If file names are given years are manually selected now
Before when passing explicit file list to open, no year selection was made because it was assumed that only files
were passed that actually correspond to the correct years. But if there
is for example a file that in concatenated over multiple years, no year
selection was made inside this function and thus monthly means had to be
calculated outside after manually selecting years. Now, inside the
function the year variable is used to select years (and months, day,
record, as previously) and thus time arithmetic can be performed as
usual inside this function